Moved confidence calculation to separate function
